chore: make bytecode & abi packing optional in pack.sh

+++ pack.sh
@@ -30,6 +30,8 @@ do
   cat $DIST_SRC_DIR$CONTRACT_NAME.sol | jq -Rs . > $DIST_SRC_DIR$CONTRACT_NAME.json
   rm $DIST_SRC_DIR$CONTRACT_NAME.sol
 
-  cat ./out/$CONTRACT_NAME.sol/$CONTRACT_NAME.json | jq '{bytecode: .bytecode.object, linkReferences: .bytecode.linkReferences}' > $DIST_BIN_DIR$CONTRACT_NAME.json
-  cat ./out/$CONTRACT_NAME.sol/$CONTRACT_NAME.json | jq '.abi' > $DIST_ABI_DIR$CONTRACT_NAME.json
+  if [ -f ./out/$CONTRACT_NAME.sol/$CONTRACT_NAME.json ]; then
+    cat ./out/$CONTRACT_NAME.sol/$CONTRACT_NAME.json | jq '{bytecode: .bytecode.object, linkReferences: .bytecode.linkReferences}' > $DIST_BIN_DIR$CONTRACT_NAME.json
+    cat ./out/$CONTRACT_NAME.sol/$CONTRACT_NAME.json | jq '.abi' > $DIST_ABI_DIR$CONTRACT_NAME.json
+  fi
 done
